Initial PR review fixes
